requirements.txt
Django==2.2.6
pytz==2019.3
sqlparse==0.3.0
mysqlclient==1.4.4
djangorestframework==3.10.3